Moved router into separate router.ts file
